+++ template-parts/menu/menu-social.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0.0
- * @version  2.0.0
+ * @version  2.2.3
  */
 
 
@@ -20,19 +20,19 @@
 	}
 
 
-// Helper variables
+// Variables
 
-	$social_menu_html = get_transient( 'modern_social_links' );
-	$social_menu_args = array(
-		'theme_location' => 'social',
-		'container'      => false,
-		'menu_class'     => 'social-links-items',
-		'depth'          => 1,
-		'link_before'    => '<span class="screen-reader-text">',
-		'link_after'     => '</span>',
-		'fallback_cb'    => false,
-		'items_wrap'     => '<ul data-id="%1$s" class="%2$s">%3$s<li class="back-to-top-link"><a href="#" class="back-to-top" title="' . esc_attr__( 'Back to top', 'modern' ) . '"><span class="screen-reader-text">' . esc_html__( 'Back to top &uarr;', 'modern' ) . '</span></a></li></ul>',
-	);
+	$cache_key   = Modern_Menu::get_cache_key_social();
+	$cache_group = 'modern_' . get_bloginfo( 'language' );
+
+	$back_to_top  = '<li class="back-to-top-link">';
+	$back_to_top .= '<a href="#page" class="back-to-top animated" title="' . esc_attr__( 'Back to top', 'modern' ) . '">';
+	$back_to_top .= '<span class="screen-reader-text">' . esc_html__( 'Back to top &uarr;', 'modern' ) . '</span>';
+	$back_to_top .= '</a>';
+	$back_to_top .= '</li>';
+
+	$social_menu_html = wp_cache_get( $cache_key, $cache_group );
+	$social_menu_args = Modern_Menu::get_menu_args_social( '<ul data-id="%1$s" class="%2$s">%3$s' . $back_to_top . '</ul>' );
 
 
 ?>
@@ -45,6 +45,8 @@
 
 		/**
 		 * If we want to enable customizer partial edit, we need to output the menu standard way.
+		 *
+		 * @subpackage  Customize
 		 */
 		wp_nav_menu( $social_menu_args );
 
@@ -55,10 +57,22 @@
 		 *
 		 * @see  Modern_Menu::social_cache_flush()
 		 */
-		if ( ! $social_menu_html ) {
-			$social_menu_html = wp_nav_menu( array_merge( array( 'echo' => false ), $social_menu_args ) );
-			$social_menu_html = str_replace( ' id=', ' data-id=', $social_menu_html ); // Fix for multiple displays
-			set_transient( 'modern_social_links', $social_menu_html );
+		if ( empty( $social_menu_html ) ) {
+			$social_menu_args['echo'] = false;
+
+			$social_menu_html = wp_nav_menu( $social_menu_args );
+			$social_menu_html = str_replace(
+				' id=',
+				' data-id=',
+				$social_menu_html
+			);
+
+			wp_cache_set(
+				$cache_key,
+				$social_menu_html,
+				$cache_group,
+				7 * 24 * 60 * 60
+			);
 		}
 
 		echo $social_menu_html;